Linux CM Kick-off (#186)
* Initial commit on ARM, need to move to Intel to continue

* Advanced associated submodules

* More changes to allow us to build on Linux

* It's building, yay

Documentation added

adjusting the Secure Client UCB installer

Some clean up

Change token

add missing component to add dependencies

Try different way of quoting

Add missing common path due to refactoring

Fix quoting

clean up

Removing redundant lines

Update gitmodules to reflect submodule fork

Test - Add Linux CI

Fix syntax error

Add missing import

Add missing method

More fun

Fix

Imitate ampcx development jenkinsfle

Trying things

Try try try

Fix parameter

test

experiment

Test test test

Move back to using env vars directly

Set Linux true when not APPLE

Missing brace

Fix references to paths on the build machine

Fix exclude paths

Advance EndpointIdentity

Moving Mac over to ephemeral nodes as well

More clean up

* Try Linux with devtoolset

* advance EndpointIdentity

* help with linkage errors

* Try 2

* Additional build fixes

* Re-enable tests

* Advance package manager

* Advance PackageManager to fix Mac build

* Advance EndpointIdentity

* Advance EndpointIdentity submodule again

* Add support for clean workspace label

* Disable Ubuntu add the rest of the Mac stages back in

* Adding back continueCI but it can probably be removed

* Fix release builds

* Advance Package Manager

* Linking test again

* Clean up before starting xcode build

* Advancing EndpointIdentitiy

* Add a clean step

* clean up and address comments

* clean for Xcode release build should not be necessary

* Advancing EndpointIdentity

* Advance PackageManager, and gtest after merged of changes to upstream

* Advance EndpointIdentity to point to upstream master

+++ scripts/build.sh
@@ -1,4 +1,4 @@
-#!/bin/sh -e
+#!/bin/bash -e
 # Copyright 2022 Cisco Systems, Inc.
 
 # set -x
@@ -228,6 +228,7 @@ else
             else
                 cmake ${CMAKE_EXTRA_ARGS} ../
             fi
+	    # Optionally add --verbose for debugging build issues
             cmake --build .
         popd
 
@@ -246,27 +247,29 @@ else
                 cmake --install .
             popd
         
-            echo "** Building CM Installer **"
-            BUILD_TYPE="${CMAKE_BUILD_DIR}"
-            DMG_INSTALLER_DIR="Installer"
-            DMG_BUILDER_SCRIPT="build_cm_installer.sh"
+            if [ "${SYSTEM}" = "Darwin" ]; then 
+                echo "** Building CM Installer **"
+                BUILD_TYPE="${CMAKE_BUILD_DIR}"
+                DMG_INSTALLER_DIR="Installer"
+                DMG_BUILDER_SCRIPT="build_cm_installer.sh"
 
-            if [ -d "${BUILD_STAGING_DIR}" ]; then
-                echo "Removing existing build Staging directory"
-                rm -rf "${BUILD_STAGING_DIR}"
+                if [ -d "${BUILD_STAGING_DIR}" ]; then
+                    echo "Removing existing build Staging directory"
+                    rm -rf "${BUILD_STAGING_DIR}"
+                fi
+
+                pushd "${DMG_INSTALLER_DIR}"
+                    "./${DMG_BUILDER_SCRIPT}" "${BUILD_TYPE}" "${BUILD_STAGING_DIR}"
+                    echo "** CM Installer built successfully **"
+                popd
+
+                echo "** Building Sample Installer **"
+                SAMPLE_PKG_BUILDER_SCRIPT="build_sample_installer.sh"
+                pushd "${DMG_INSTALLER_DIR}"
+                    "./${SAMPLE_PKG_BUILDER_SCRIPT}" "${BUILD_TYPE}" "${BUILD_STAGING_DIR}"
+                    echo "** Sample Installer built successfully **"
+                popd
             fi
-
-            pushd "${DMG_INSTALLER_DIR}"
-                "./${DMG_BUILDER_SCRIPT}" "${BUILD_TYPE}" "${BUILD_STAGING_DIR}"
-                echo "** CM Installer built successfully **"
-            popd
-
-            echo "** Building Sample Installer **"
-            SAMPLE_PKG_BUILDER_SCRIPT="build_sample_installer.sh"
-            pushd "${DMG_INSTALLER_DIR}"
-                "./${SAMPLE_PKG_BUILDER_SCRIPT}" "${BUILD_TYPE}" "${BUILD_STAGING_DIR}"
-                echo "** Sample Installer built successfully **"
-            popd
 
         fi
     fi
